+++ benches/ring.rs
@@ -3,7 +3,7 @@ mod bench_utils;
 
 use ark_std::UniformRand;
 use ark_vrf::{
-    AffinePoint, Input, Secret, VrfIo,
+    AffinePoint, Input, Public, Secret, VrfIo,
     ring::{self, BatchItem, BatchVerifier, Proof, RingSuite},
 };
 use bench_utils::SuiteExt;
@@ -15,7 +15,7 @@ const RING_SIZES: [usize; 3] = [255, 1023, 2047];
 struct BenchSetup<S: RingSuite> {
     secret: Secret<S>,
     io: VrfIo<S>,
-    ring: Vec<AffinePoint<S>>,
+    ring: Vec<Public<S>>,
     prover_idx: usize,
     ring_setup: ring::RingSetup<S>,
 }
@@ -28,10 +28,10 @@ fn make_ring_setup<S: RingSuite>(ring_size: usize) -> BenchSetup<S> {
     let io = secret.vrf_io(input);
 
     let prover_idx = 3;
-    let mut ring: Vec<AffinePoint<S>> = (0..ring_size)
-        .map(|_| AffinePoint::<S>::rand(&mut rng))
+    let mut ring: Vec<Public<S>> = (0..ring_size)
+        .map(|_| Public::from_affine_unchecked(AffinePoint::<S>::rand(&mut rng)))
         .collect();
-    ring[prover_idx] = public.point();
+    ring[prover_idx] = public;
 
     let ring_setup = ring::RingSetup::<S>::from_rand_insecure(ring_size, &mut rng);
 
@@ -91,7 +91,7 @@ fn ring_benches<S: RingSuite>(c: &mut Criterion) {
             });
 
         let ring_ctx = setup.ring_setup.ring_context();
-        let prover_key = setup.ring_setup.prover_key(&setup.ring).unwrap();
+        let (prover_key, verifier_key) = setup.ring_setup.keys(&setup.ring).unwrap();
         let prover = ring_ctx.ring_prover(prover_key, setup.prover_idx);
 
         c.benchmark_group(format!("{}/ring_prove", S::SUITE_NAME))
@@ -101,7 +101,6 @@ fn ring_benches<S: RingSuite>(c: &mut Criterion) {
             });
 
         let proof = Proof::prove(setup.io, b"ad", &setup.secret, &prover);
-        let verifier_key = setup.ring_setup.verifier_key(&setup.ring).unwrap();
         let commitment = verifier_key.commitment();
         let verifier = ring_ctx.ring_verifier(verifier_key.clone());
 
@@ -172,7 +171,7 @@ fn batch_benches<S: RingSuite>(c: &mut Criterion) {
     let setup = make_ring_setup::<S>(1023);
 
     let ring_ctx = setup.ring_setup.ring_context();
-    let prover_key = setup.ring_setup.prover_key(&setup.ring).unwrap();
+    let (prover_key, verifier_key) = setup.ring_setup.keys(&setup.ring).unwrap();
     let prover = ring_ctx.ring_prover(prover_key, setup.prover_idx);
 
     let max_batch_size = BATCH_SIZES[BATCH_SIZES.len() - 1];
@@ -196,7 +195,6 @@ fn batch_benches<S: RingSuite>(c: &mut Criterion) {
         })
         .collect();
 
-    let verifier_key = setup.ring_setup.verifier_key(&setup.ring).unwrap();
     let verifier = ring_ctx.ring_verifier(verifier_key);
 
     // batch_verifier_new: cost is independent of batch size, bench once.